tools/document_ai/requirements.txt
pdfplumber>=0.10.0
pytesseract>=0.3.10
pydantic>=2.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
requests>=2.31.0
